More SOLR work - almost got integration test going

+++ discodj-core/src/main/java/com/totalchange/discodj/search/SearchFacet.java
@@ -3,5 +3,5 @@ package com.totalchange.discodj.search;
 public interface SearchFacet {
     String getId();
     String getName();
-    int getNumMatches();
+    long getNumMatches();
 }
